Allow RegEx in the RF Network Monitor
Allows the use of regular expressions in the RF Network Monitor to give users who are familiar with regular expressions more power in their filters. This should not affect users who do not know RegEx except in rare cases (when regex special characters are unintentionally used, which is extremely rare, especially in Minecraft).